requirements.txt
Flask==3.0.0
openpyxl==3.1.2
gunicorn==21.2.0
gspread==5.12.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
